Pom: ajout de junit, ca peut tjr etre utile
Point: suppression fonction inutile

area: fix recherche par id et changement du chemin
area: ajout fonction de delete

+++ src/main/java/application/repository/AreaRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface AreaRepository extends CrudRepository<Area, Long> {
 
-    Area findById(long id);
+    Area findById(Integer id);
     Area findByName(String name);
 //    List<Area> findByCoordinates(String coordinates);
